+++ AcademyCourses/AcademyCourses/ModuloBE.cs
@@ -10,13 +10,13 @@ namespace AcademyCourses
 
         public int C_Modulo { get; set; }
         public CategoriaBE C_Categoria { get; set; }
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
         public String Descripcion { get; set; }
         public Boolean Estado { get; set; }
 
         public ModuloBE () { }
 
-        public ModuloBE(int C_Modulo, CategoriaBE C_Categoria, double Precio, String Descripcion, Boolean Estado) {
+        public ModuloBE(int C_Modulo, CategoriaBE C_Categoria, decimal Precio, String Descripcion, Boolean Estado) {
 
             this.C_Modulo = C_Modulo;
             this.C_Categoria = C_Categoria;
